+++ examples/auth-google-login/src/App.tsx
@@ -101,7 +101,10 @@ const App: React.FC = () => {
 
             return {
                 authenticated: false,
-                error: new Error("Not authenticated"),
+                error: {
+                    message: "Check failed",
+                    name: "Not authenticated",
+                },
                 logout: true,
                 redirectTo: "/login",
             };