+++ examples/data-provider-appwrite/src/App.tsx
@@ -88,7 +88,10 @@ const authProvider: AuthBindings = {
 
         return {
             authenticated: false,
-            error: new Error("Session not found"),
+            error: {
+                message: "Check failed",
+                name: "Session not found",
+            },
             logout: true,
             redirectTo: "/login",
         };