+++ examples/data-provider-supabase/src/App.tsx
@@ -78,7 +78,10 @@ const authProvider: AuthBindings = {
 
         return {
             success: false,
-            error: new Error("Login failed"),
+            error: {
+                message: "Login failed",
+                name: "Invalid email or password",
+            },
         };
     },
     register: async ({ email, password }) => {
@@ -109,7 +112,10 @@ const authProvider: AuthBindings = {
 
         return {
             success: false,
-            error: new Error("Register failed"),
+            error: {
+                message: "Register failed",
+                name: "Invalid email or password",
+            },
         };
     },
     forgotPassword: async ({ email }) => {
@@ -146,7 +152,10 @@ const authProvider: AuthBindings = {
 
         return {
             success: false,
-            error: new Error("Forgot Password password failed"),
+            error: {
+                message: "Forgot password failed",
+                name: "Invalid email",
+            },
         };
     },
     updatePassword: async ({ password }) => {
@@ -176,7 +185,10 @@ const authProvider: AuthBindings = {
         }
         return {
             success: false,
-            error: new Error("Update Password password failed"),
+            error: {
+                message: "Update password failed",
+                name: "Invalid password",
+            },
         };
     },
     logout: async () => {
@@ -206,7 +218,10 @@ const authProvider: AuthBindings = {
             if (!session) {
                 return {
                     authenticated: false,
-                    error: new Error("Not authenticated"),
+                    error: {
+                        message: "Check failed",
+                        name: "Session not found",
+                    },
                     logout: true,
                     redirectTo: "/login",
                 };
@@ -214,7 +229,10 @@ const authProvider: AuthBindings = {
         } catch (error: any) {
             return {
                 authenticated: false,
-                error: error || new Error("Not authenticated"),
+                error: error || {
+                    message: "Check failed",
+                    name: "Session not found",
+                },
                 logout: true,
                 redirectTo: "/login",
             };