+++ examples/tutorial-mui/src/pages/blog-posts/edit.tsx
@@ -91,7 +91,8 @@ export const BlogPostEdit = () => {
                             }}
                             isOptionEqualToValue={(option, value) =>
                                 value === undefined ||
-                                option?.id?.toString() === value?.id?.toString()
+                                option?.id?.toString() ===
+                                    (value?.id ?? value)?.toString()
                             }
                             renderInput={(params) => (
                                 <TextField