+++ examples/with-nextjs/src/authProvider.ts
@@ -58,7 +58,10 @@ export const authProvider: AuthBindings = {
               }
             : {
                   authenticated: false,
-                  error: new Error("Unauthorized"),
+                  error: {
+                      message: "Check failed",
+                      name: "Unauthorized",
+                  },
                   logout: true,
                   redirectTo: "/login",
               };