+++ packages/core/src/hooks/router/use-parsed/index.tsx
@@ -1,10 +1,12 @@
 import React from "react";
 import { useParse } from "../use-parse";
 
-export const useParsed = () => {
+export const useParsed = <
+    TParams extends Record<string, any> = Record<string, any>,
+>() => {
     const parse = useParse();
 
-    const parsed = React.useMemo(() => parse(), [parse]);
+    const parsed = React.useMemo(() => parse<TParams>(), [parse]);
 
     return parsed;
 };