+++ packages/core/src/interfaces/errors/index.ts
@@ -1,0 +1,2 @@
+export * from "./HttpError";
+export * from "./RefineError";